+++ controllers/HartaController.js
@@ -1,4 +1,5 @@
 import Harta from "../models/HartaModel.js";
+import { createError } from "../utils/error.js";
 
 export const getHartas = async (req, res) => {
   try {
@@ -43,17 +44,17 @@ export const getHartaLancarAll = async (req, res) => {
 
 export const getHartaLancarAllForDoc = async (req, res) => {
   try {
-    let tempHarta = []
+    let tempHarta = [];
     const harta = await Harta.find({}).sort({ createdAt: -1 }).limit(1);
     const hartaLancar = harta[0].hartaLancar;
-    hartaLancar.map(val => {
+    hartaLancar.map((val) => {
       tempHarta.push({
         kodeAccount: val.kodeAccount,
         namaAccount: val.namaAccount,
         kelompokAccount: val.kelompokAccount,
         total: val.total.toLocaleString(),
-      })
-    })
+      });
+    });
     res.json(tempHarta);
   } catch (error) {
     // Error 500 = Kesalahan di server
@@ -106,17 +107,17 @@ export const getHartaTetapAll = async (req, res) => {
 
 export const getHartaTetapAllForDoc = async (req, res) => {
   try {
-    let tempHarta = []
+    let tempHarta = [];
     const harta = await Harta.find({}).sort({ createdAt: -1 }).limit(1);
     const hartaTetap = harta[0].hartaTetap;
-    hartaTetap.map(val => {
+    hartaTetap.map((val) => {
       tempHarta.push({
         kodeAccount: val.kodeAccount,
         namaAccount: val.namaAccount,
         kelompokAccount: val.kelompokAccount,
         total: val.total.toLocaleString(),
-      })
-    })
+      });
+    });
     res.json(tempHarta);
   } catch (error) {
     // Error 500 = Kesalahan di server
@@ -156,42 +157,54 @@ export const getHartaTetapOther = async (req, res) => {
   }
 };
 
-export const saveHarta = async (req, res) => {
+export const saveHarta = async (req, res, next) => {
   const harta = new Harta(req.body);
   try {
-    const insertedHarta = await harta.save();
-    // Status 201 = Created
-    res.status(201).json(insertedHarta);
+    if (req.user.isAdmin) {
+      const insertedHarta = await harta.save();
+      // Status 201 = Created
+      res.status(201).json(insertedHarta);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
   }
 };
 
-export const updateHarta = async (req, res) => {
+export const updateHarta = async (req, res, next) => {
   try {
-    const updatedHarta = await Harta.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    // Status 200 = Successful
-    res.status(200).json(updatedHarta);
+    if (req.user.isAdmin) {
+      const updatedHarta = await Harta.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: req.body,
+        },
+        { new: true }
+      );
+      // Status 200 = Successful
+      res.status(200).json(updatedHarta);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
   }
 };
 
-export const deleteHarta = async (req, res) => {
+export const deleteHarta = async (req, res, next) => {
   try {
-    const deletedHarta = await Harta.deleteOne({
-      _id: req.params.id,
-    });
-    // Status 200 = Successful
-    res.status(200).json(deletedHarta);
+    if (req.user.isAdmin) {
+      const deletedHarta = await Harta.deleteOne({
+        _id: req.params.id,
+      });
+      // Status 200 = Successful
+      res.status(200).json(deletedHarta);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
