+++ controllers/NeracaController.js
@@ -1,4 +1,5 @@
 import Neraca from "../models/NeracaModel.js";
+import { createError } from "../utils/error.js";
 
 export const getNeracas = async (req, res) => {
   try {
@@ -20,42 +21,54 @@ export const getNeracaLast = async (req, res) => {
   }
 };
 
-export const saveNeraca = async (req, res) => {
+export const saveNeraca = async (req, res, next) => {
   const neraca = new Neraca(req.body);
   try {
-    const insertedNeraca = await neraca.save();
-    // Status 201 = Created
-    res.status(201).json(insertedNeraca);
+    if (req.user.isAdmin) {
+      const insertedNeraca = await neraca.save();
+      // Status 201 = Created
+      res.status(201).json(insertedNeraca);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
   }
 };
 
-export const updateNeraca = async (req, res) => {
+export const updateNeraca = async (req, res, next) => {
   try {
-    const updatedNeraca = await Neraca.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    // Status 200 = Successful
-    res.status(200).json(updatedNeraca);
+    if (req.user.isAdmin) {
+      const updatedNeraca = await Neraca.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: req.body,
+        },
+        { new: true }
+      );
+      // Status 200 = Successful
+      res.status(200).json(updatedNeraca);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
   }
 };
 
-export const deleteNeraca = async (req, res) => {
+export const deleteNeraca = async (req, res, next) => {
   try {
-    const deletedNeraca = await Neraca.deleteOne({
-      _id: req.params.id,
-    });
-    // Status 200 = Successful
-    res.status(200).json(deletedNeraca);
+    if (req.user.isAdmin) {
+      const deletedNeraca = await Neraca.deleteOne({
+        _id: req.params.id,
+      });
+      // Status 200 = Successful
+      res.status(200).json(deletedNeraca);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
