+++ controllers/PerubahanModalController.js
@@ -1,4 +1,5 @@
 import PerubahanModal from "../models/PerubahanModalModel.js";
+import { createError } from "../utils/error.js";
 
 export const getPerubahanModals = async (req, res) => {
   try {
@@ -12,7 +13,7 @@ export const getPerubahanModals = async (req, res) => {
 
 export const getPerubahanModalForDoc = async (req, res) => {
   try {
-    let tempPerubahanModal = []
+    let tempPerubahanModal = [];
     const perubahanModal = await PerubahanModal.find(
       {},
       {
@@ -22,13 +23,13 @@ export const getPerubahanModalForDoc = async (req, res) => {
         _id: 0,
       }
     );
-    perubahanModal.map(val => {
+    perubahanModal.map((val) => {
       tempPerubahanModal.push({
         modalSaham: val.modalSaham.toLocaleString(),
         labaBersih: val.labaBersih.toLocaleString(),
         total: val.total.toLocaleString(),
-      })
-    })
+      });
+    });
     res.json(tempPerubahanModal);
   } catch (error) {
     // Error 500 = Kesalahan di server
@@ -48,42 +49,54 @@ export const getPerubahanModalLast = async (req, res) => {
   }
 };
 
-export const savePerubahanModal = async (req, res) => {
+export const savePerubahanModal = async (req, res, next) => {
   const perubahanModal = new PerubahanModal(req.body);
   try {
-    const insertedPerubahanModal = await perubahanModal.save();
-    // Status 201 = Created
-    res.status(201).json(insertedPerubahanModal);
+    if (req.user.isAdmin) {
+      const insertedPerubahanModal = await perubahanModal.save();
+      // Status 201 = Created
+      res.status(201).json(insertedPerubahanModal);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
   }
 };
 
-export const updatePerubahanModal = async (req, res) => {
+export const updatePerubahanModal = async (req, res, next) => {
   try {
-    const updatedPerubahanModal = await PerubahanModal.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    // Status 200 = Successful
-    res.status(200).json(updatedPerubahanModal);
+    if (req.user.isAdmin) {
+      const updatedPerubahanModal = await PerubahanModal.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: req.body,
+        },
+        { new: true }
+      );
+      // Status 200 = Successful
+      res.status(200).json(updatedPerubahanModal);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
   }
 };
 
-export const deletePerubahanModal = async (req, res) => {
+export const deletePerubahanModal = async (req, res, next) => {
   try {
-    const deletedPerubahanModal = await PerubahanModal.deleteOne({
-      _id: req.params.id,
-    });
-    // Status 200 = Successful
-    res.status(200).json(deletedPerubahanModal);
+    if (req.user.isAdmin) {
+      const deletedPerubahanModal = await PerubahanModal.deleteOne({
+        _id: req.params.id,
+      });
+      // Status 200 = Successful
+      res.status(200).json(deletedPerubahanModal);
+    } else {
+      return next(createError(403, "You are not authorized!"));
+    }
   } catch (error) {
     // Error 400 = Kesalahan dari sisi user
     res.status(400).json({ message: error.message });
